Split runtime backend and executable source files (#2544)
* split runtime backend and executable source files

* style
