+++ src/ngraph/runtime/backend.hpp
@@ -20,6 +20,7 @@
 
 #include "ngraph/function.hpp"
 #include "ngraph/pass/pass_config.hpp"
+#include "ngraph/runtime/executable.hpp"
 #include "ngraph/runtime/performance_counter.hpp"
 #include "ngraph/shape.hpp"
 #include "ngraph/type/element_type.hpp"
@@ -28,10 +29,8 @@ namespace ngraph
 {
     namespace runtime
     {
-        class ExternalFunction;
         class Tensor;
         class Backend;
-        class Executable;
     }
 }
 
@@ -110,52 +109,4 @@ public:
     virtual bool is_supported_property(const Property prop) const;
 
     virtual void remove_compiled_function(std::shared_ptr<Executable> exec);
-};
-
-class ngraph::runtime::Executable
-{
-public:
-    Executable();
-    virtual ~Executable();
-
-    /// \param outputs vector of runtime::Tensor used as outputs
-    /// \param inputs vector of runtime::Tensor used as inputs
-    /// \returns true if iteration is successful, false otherwise
-    virtual bool call(const std::vector<std::shared_ptr<runtime::Tensor>>& outputs,
-                      const std::vector<std::shared_ptr<runtime::Tensor>>& inputs) = 0;
-
-    /// \brief Executes a single iteration of a Function.
-    /// \param outputs vector of runtime::Tensor used as outputs
-    /// \param inputs vector of runtime::Tensor used as inputs
-    /// \returns true if iteration is successful, false otherwise
-    bool call_with_validate(const std::vector<std::shared_ptr<runtime::Tensor>>& outputs,
-                            const std::vector<std::shared_ptr<runtime::Tensor>>& inputs);
-
-    /// \brief Collect performance information gathered on a Function.
-    /// \returns Vector of PerformanceCounter information.
-    virtual std::vector<PerformanceCounter> get_performance_data() const;
-
-    /// \brief Validates a Function.
-    /// \param outputs vector of runtime::Tensor used as outputs
-    /// \param inputs vector of runtime::Tensor used as inputs
-    void validate(const std::vector<std::shared_ptr<runtime::Tensor>>& outputs,
-                  const std::vector<std::shared_ptr<runtime::Tensor>>& inputs);
-
-    /// \brief Query the input Parameters
-    /// \returns an ngraph::op::ParameterVector of all input parameters
-    const ngraph::ParameterVector& get_parameters() const;
-
-    /// \brief Query the output Results
-    /// \returns an ngraph::ResultVector of all input parameters
-    const ngraph::ResultVector& get_results() const;
-
-protected:
-    /// \brief Called at the end of compile to the values to be returned by get_parameters
-    ///     and get_results
-    /// \param func The function with Results fully resolved.
-    void set_parameters_and_results(const Function& func);
-
-private:
-    ngraph::ParameterVector m_parameters;
-    ngraph::ResultVector m_results;
 };
